Creates publication-ready summary tables

Creates publication-ready summary tables based on the gtsummary package.

summaryTable(
  data,
  vars = NULL,
  group = NULL,
  labels = NULL,
  stat_cont = "median_range",
  stat_cat = "n_percent",
  test = FALSE,
  test_cont = "wilcox.test",
  test_cat = "chisq.test",
  ci = FALSE,
  ci_cont = "wilcox.test",
  ci_cat = "wilson",
  conf_level = 0.95,
  digits_cont = 1,
  digits_cat = 0,
  missing = TRUE,
  binary = FALSE,
  missing_text = "Missing",
  overall = FALSE,
  as_flex_table = TRUE
)

Arguments

data

A data frame or tibble containing the data to be summarized.

vars

Variables to include in the summary table. Default to all variables present in the data except group.

group

A single column from data. Summary statistics will be stratified according to this variable. Default to NULL.

labels

A list containing the labels that should be used for the variables in the table. If NULL, labels are automatically taken from the dataset. If no label present, the variable name is taken.

stat_cont

Summary statistic to display for continuous variables. Options include "median_IQR", "median_range" (default), "mean_sd", "mean_se" and "geomMean_sd".

stat_cat

Summary statistic to display for categorical variables. Options include "n_percent" (default) and "n", and "n_N".

test

Logical. Indicates whether p-values are displayed (TRUE) or not (FALSE). Default to FALSE

test_cont

Test type used to calculate the p-value for continuous variables. Only used if test = TRUE. Options include "t.test", "oneway.test", "kruskal.test", "wilcox.test" (default), "paired.t.test", "paired.wilcox.test"

test_cat

Test type used to calculated the p-value for categorical variables. Only used if test = TRUE. Options include "chisq.test" (default), "chisq.test.no.correct", "fisher.test".

ci

Logical. Indicates whether CI are displayed (TRUE) or not (FALSE). Default to FALSE.

ci_cont

Confidence interval method for continuous variables. Only used if ci = TRUE. Options include "t.test" and "wilcox.test" (default).

ci_cat

Confidence interval method for categorical variables. Options include "wilson" (default), "wilson.no.correct", "clopper.pearson", "wald", "wald.no.correct", "agresti.coull" and "jeffreys". If NULL, no CI will be displayed.

conf_level

Confidence level. Default to 0.95.

digits_cont

Digits for summary statistics and CI of continuous variables. Default to 1.

digits_cat

Digits for summary statistics and CI of categorical variables. Default to 0.

missing

Indicates whether percentages for missings are shown (TRUE, default) or not (FALSE) for categorical variables.Categorical variables need to be factors for it to work. If "both", then both options are displayed next to each other.

binary

Only taken into account when missing = FALSE. For binary variables, indicates whether all levels are displayed (FALSE, default) or only one.

missing_text

String indicating text shown on missing row. Default to "Missing".

overall

Logical. If TRUE, an additional column with the total is added to the table. Default to FALSE.

as_flex_table

Logical. If TRUE (default) the gtsummary object is converted to a flextable object. Useful when rendering to Word.

Examples

library(SAKK)
library(survival)
data("cancer")
summaryTable(data = cancer,vars = c("inst", "time","age", "ph.ecog"),
             labels = list(inst = "Institution code",
                           time = "Time",
                           age = "Age",
                           ph.ecog = "ECOG score"))

Characteristic

N = 2281

Institution code

11.0 (1.0, 33.0)

Missing

1

Time

255.5 (5.0, 1,022.0)

Age

63.0 (39.0, 82.0)

ECOG score

0

63 (28%)

1

113 (50%)

2

50 (22%)

3

1 (0%)

Missing

1 (0%)

1Median (Min, Max); n (%)